fix(editor): some bugs of member column (#11033)
fix: BS-2840
